requirements.txt
beautifulsoup4==4.12.3
faiss-cpu==1.8.0
langchain==0.1.10
langchain-openai==0.0.8
python-dotenv==1.0.1
xmltodict==0.13.0
